+++ app/src/main/java/com/example/finalapp/featureNote/domain/useCases/GetByIdItemUseCase.kt
@@ -5,7 +5,7 @@ import com.example.finalapp.featureNote.domain.repositories.ItemRepository
 
 class GetByIdItemUseCase(private val repository: ItemRepository) {
 
-    suspend operator fun invoke(itemId: Long): Item? {
+    suspend operator fun invoke(itemId: Int): Item? {
         return repository.getItemById(itemId)
     }
 }
